[S14P31S309-1301] feat: 키보드 숫자/특수문자 모드 추가

See merge request s14-final/S14P31S309!284

+++ frontend/app/src/main/java/com/ssafy/s309/data/repository/source/SamsungHealthDataSource.kt
@@ -39,10 +39,10 @@ class SamsungHealthDataSource
                 val steps = mgr.getTodaySteps().toInt()
                 val sleep = mgr.getLastSleepDurationMinutes()
                 // 권한 미부여 / 데이터 없음 시 둘 다 0 → 데이터 없음으로 간주 → mock fallback
-                if (steps == 0 && sleep == 0) {
+                if (steps == 0L && sleep == 0) {
                     null
                 } else {
-                    DailyHealthSummary(steps = steps, sleepMinutes = sleep)
+                    DailyHealthSummary(steps = steps.toInt(), sleepMinutes = sleep)
                 }
             } catch (t: Throwable) {
                 Log.w(TAG, "Samsung Health summary fetch 실패", t)
